factor out renderCategory for re-use
